test(rbac,geoip): move doc.go Examples into example_test.go so go test actually runs them
func Example in a non-_test.go file is never compiled or executed by go test — only _test.go files are scanned for examples — so both packages' doctests could silently rot. Moved verbatim into black-box example_test.go files (package X_test, exported-API calls prefixed); doc.go prose untouched. Both examples now show up as RUN/PASS in go test -run Example -v.
