+++ web/geoip/doc.go
@@ -18,22 +18,3 @@
 //
 //	logger.New(logger.WithContextExtractors(geoip.LogExtractor))
 package geoip
-
-import (
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-)
-
-// Example wires a header-only source (no database file needed) and reads the
-// cached Location in a handler.
-func Example() {
-	src := Cloudflare()
-	h := Middleware(src)(http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-		fmt.Println(Get(r).CountryCode)
-	}))
-	r := httptest.NewRequest(http.MethodGet, "/", nil)
-	r.Header.Set("CF-IPCountry", "US")
-	h.ServeHTTP(httptest.NewRecorder(), r)
-	// Output: US
-}
